feat: clean calendar code and fix spacing
