+++ apps/website/src/components/datepicker/datepicker.tsx
@@ -16,7 +16,7 @@ const DatePicker = (props: Props) => {
 
   return (
     <div className="sticky bottom-5 flex justify-center">
-      <Popover>
+      <Popover alignOffset={8} align="center" sideOffset={8}>
         <button className="border-neutral-80/5 bg-blur-neutral-5/70 active inline-flex min-h-[30px] cursor-pointer items-center justify-center gap-2 rounded-xl border-solid pl-3 pr-2 uppercase text-neutral-100 backdrop-blur-sm">
           <span className="text-blur-neutral-80/80 text-[13px] font-medium">
             Filter between
@@ -30,7 +30,12 @@ const DatePicker = (props: Props) => {
           <EditIcon size={20} color="$neutral-80-opa-40" />
         </button>
         <Popover.Content>
-          <Calendar mode="range" selected={selected} onSelect={onSelect} />
+          <Calendar
+            mode="range"
+            selected={selected}
+            onSelect={onSelect}
+            fixedWeeks
+          />
         </Popover.Content>
       </Popover>
     </div>
